[alias] Use bat instead of cat
Also escape cat so it is used instead of bat for loading of wal
terminal colors.
While the colorscheme is already set in kitty (the current
terminal), other terminals don't have it set if this invocation.
is removed. This therefore is kind of a fallback if I haven't
managed to or didn't have the time yet to set the wal colorscheme
in other terminals.